pytest_otel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytest-otel"
version = "0.1.0"
description = "Pytest plugin for OpenTelemetry instrumentation with Dagger"
readme = "README.md"
authors = [{ name = "Dagger", email = "hello@dagger.io" }]
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "pytest>=7.0.0",
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-exporter-otlp-proto-http>=1.23.0",
]
classifiers = [
    "Framework :: Pytest",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.entry-points.pytest11]
pytest_otel = "pytest_otel.plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/pytest_otel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
# Disable the pytest_otel plugin when running tests on itself
# to prevent the plugin from interfering with its own tests
addopts = "-p no:pytest_otel"

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py310"
